requirements.txt
beautifulsoup4>=4.12
duckdb>=0.10
fastapi>=0.111
geopandas>=0.14
osmnx>=1.9
pandas>=2.1
paho-mqtt>=2.0
plotly>=5.20
pyarrow>=15
pydantic>=2.6
pydeck>=0.8
pyproj>=3.6
pyyaml>=6.0
requests>=2.31
ruff>=0.4
shapely>=2.0
streamlit>=1.33
uvicorn[standard]>=0.29
pytest>=8.0
